fix(pallet-chainlink): tests of pallet-chainlink
