+++ pallet-chainlink/src/lib.rs
@@ -19,6 +19,8 @@
 pub use pallet::*;
 
 #[cfg(test)]
+mod mock;
+#[cfg(test)]
 mod tests;
 
 #[frame_support::pallet]
